backend/python-server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
aiofiles==23.2.1
python-multipart==0.0.6
python-dotenv==1.0.0
google-adk==0.1.0
google-genai==0.3.0
manim==0.18.0
pathlib2==2.3.7
